BCS 10.9.2-SNAPSHOT release, CodecMigration System added, some improvements and bug fixes.

+++ bukkit-api/src/main/java/eu/mcone/coresystem/api/bukkit/npc/capture/codecs/ItemSwitchEventCodec.java
@@ -28,7 +28,7 @@ public class ItemSwitchEventCodec extends Codec<PlayerItemHeldEvent, PlayerNpc> 
     private String enchantments = "";
 
     public ItemSwitchEventCodec() {
-        super((byte) 4, (byte) 2);
+        super((short) 4, (short) 2);
     }
 
     @Override
